Fix usage of checking a commit range
When checking a commit range, the subcommand "commits" as well
as the first parameter "since" are marked as being optional in
the usage section but actually they aren't. Change the usage to
make it clear which arguments are optional and which are not.

These arguments were probably optional because the "check" command
can be run without any arguments to check "everything". Add a new
entry in the usage section for this usage of "check".

While at there, fix style of optional arguments in the usage
section according to git.git.

Signed-off-by: Ralf Thielow <ralf.thielow@gmail.com>